+++ include/data_structs/MolpherAtom.hpp
@@ -20,13 +20,13 @@ private:
 public:
 	enum LockingMask {
 		UNLOCKED = 0 // everything goes (the default option)
-		, NO_MUTATION = 1<<1 // always the same element // TODO: this should be set if the atom cannot be changed because of KEEP_NEIGHBORS
+		, NO_MUTATION = 1<<1 // always the same element
 		, NO_ADDITION = 1<<2 // no atoms can be added
-		, NO_REMOVAL = 1<<3 // atom cannot be removed // TODO: this should be set if the atom cannot be removed because of KEEP_NEIGHBORS
-		, KEEP_NEIGHBORS = 1<<4 // the neighboring atoms must remain (bond type can change) // TODO: this should be automatically set if KEEP_NEIGHBORS_AND_BONDS is specified
+		, NO_REMOVAL = 1<<3 // atom cannot be removed
+		, KEEP_NEIGHBORS = 1<<4 // the neighboring atoms must remain (bond type can change)
 		, KEEP_NEIGHBORS_AND_BONDS = 1<<5 // the neighboring atoms must remain (bond type cannot change)
-		//, KEEP_BONDS = 1 << 6 // keep existing bonds as they are (neighbors can change, but not the bond order) TODO: implement
-		, FULL_LOCK = NO_MUTATION | NO_ADDITION | NO_REMOVAL | KEEP_NEIGHBORS | KEEP_NEIGHBORS_AND_BONDS
+		, KEEP_BONDS = 1 << 6 // keep existing bonds as they are (neighbors can change, but not the bond order)
+		, FULL_LOCK = NO_MUTATION | NO_ADDITION | NO_REMOVAL | KEEP_NEIGHBORS | KEEP_NEIGHBORS_AND_BONDS | KEEP_BONDS
 	};
 
 	static const std::vector<LockingMask> atom_locks;