+++ tests/helpers/mol_helpers.cpp
@@ -3,6 +3,7 @@
 //
 
 #include <mol_helpers.hpp>
+#include "io/stdout.hpp"
 #include <GraphMol/ROMol.h>
 #include <GraphMol/RWMol.h>
 #include <GraphMol/SmilesParse/SmilesParse.h>
@@ -16,4 +17,15 @@ bool match_substr(std::string smiles_mol, std::string smiles_query) {
 	bool matches = RDKit::SubstructMatch(*mol, *query, match);
 
 	return matches;
+}
+
+void print_lock_info(std::shared_ptr<MolpherMol> mol) {
+	int counter = 0;
+	for (auto atm : mol->getAtoms()) {
+		std::vector<std::string> xx = MolpherAtom::lockingMaskToString(atm->getLockingMask());
+		print(atm->getSymbol() + "." + std::to_string(++counter));
+		for (auto x : xx) {
+			print(x);
+		}
+	}
 }