+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Previous version: 0.0.0b1
 - implemented a basic pathfinder for quick customized searches (`molpher.algorithms.pathfinders.BasicPathfinder`)
 - some dependencies were updated (RDKit to v2017.09.1, Boost to v1.63)
 - all dependencies are now linked dynamically and packaged separately during a conda build
-- the tests can now be built using the main CMakeLists file
+- the tests are now included in the main CMakeLists file
 - the MorphingOperator interface was implemented 
     - AddAtom operator class was added that adheres to this interface
     - RemoveAtom operator class was added that adheres to this interface
@@ -16,6 +16,7 @@ Previous version: 0.0.0b1
     - InterlayAtom operator class was added that adheres to this interface
     - ContractBond (originally BondContraction) operator class was added that adheres to this interface
     - RerouteBond (originally BondReroute) operator class was added that adheres to this interface
+- MolpherMol can now be used to get all neighbors of an atom with the getNeighbors(int idx) method
 
 ## Fixes
 - number of linked dependencies was reduced (unused libraries are not linked now)